Additional considerations for #77.
TODO: Still need to figure out how to not disrupt the current scroll position

Right now, we're using $.fn.yiiListView.update to update the viewport and reset everything. We're using this rather than deleting the node, because if we delete the node without adjusting everything, posts dissapear in the infinite-scrolling view, which causes even more issues.

Right now, the only issue with using yiiListView.update is that we can't reset the page back to the present state.

I wasn't going to do much with this until later... but since someone brought it up... ugh...